// src/app/core/models/calendar-event.model.ts
export interface EventColor {
    primary: string;
    secondary: string;
}
export interface EventAction {
    id?: string | number;
    label: string;
    cssClass?: string;
    a11yLabel?: string;
    onClick({ event, sourceEvent, }: {
        event: CalendarEvent;
        sourceEvent: MouseEvent | KeyboardEvent;
    }): any;
}
export interface CalendarEvent<MetaType = any> {
    id?: any;
    address?: any;
    city: any;
    country: any;
    start: Date;
    end?: Date;
    title: string;
    categorycolor?: any;
    color?: EventColor;
    textColor?: string;
    actions?: EventAction[];
    allDay?: boolean;
    cssClass?: string;
    timeFormat?: string;
    resizable?: {
        beforeStart?: boolean;
        afterEnd?: boolean;
    };
    draggable?: boolean;
    meta?: MetaType;
    isPublic?: any;
    hasFile: any,
    hasNote: any,
    isInvitedUserEvent?: boolean;
    isOnlineEvent?: boolean;
    isSimulation?: boolean;
    myValidationStatus?: string;
    isEventCreator?: boolean;
    isProposalEvent?: boolean;
    isExternalEvent?: boolean;
    isConvertedFromTask?:boolean;
    source?: any;
    isConsumedData?: boolean;
}